requirements.txt
# CV AI Service - Requirements
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ML / Data
scikit-learn>=1.4.0
pandas>=2.0.0
joblib>=1.3.0
numpy>=1.26.0

# PDF Reader
pdfplumber>=0.10.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0
